requirements.txt
firebase-admin==6.3.0
tweepy==4.14.0
requests==2.31.0
boto3==1.34.5
python-dotenv==1.0.0
pydantic==2.5.3
pydantic_core==2.14.6